small refactor in payload
